+++ app/views/announcements/index.html.erb
@@ -18,24 +18,11 @@
   <%= render partial: "documents/filter_form",
              locals: {
                document_type: :announcement,
-               filters: if Locale.current.english?
-                          [
-                            :keyword, :date, :announcement_type,
-                            :locations, :department,
-                            :include_world_location_news, :people, :taxon, :subtaxon
-                          ]
-                        else
-                          [ :locations ]
-                        end
+               filters: [ :locations ]
              } %>
 
   <div class="filter-results-block">
     <div class="inner-block">
-      <% if Locale.current.english? %>
-        <%= render partial: "documents/filter_results",
-                  locals: { filter: @filter,
-                            result_type: "announcement" } %>
-      <% end %>
       <%= render 'shared/feeds', atom_url: filter_atom_feed_url %>
       <div class="filter-results js-filter-results" aria-live="polite">
         <%= render_mustache('documents/filter_table', @filter.as_hash) %>
