--- app/views/documents/_filter_results.html.erb
@@ -1,28 +1,0 @@
-<%
-  context = {
-    result_count: number_with_delimiter(filter.documents.total_count),
-    pluralized_result_type: result_type.pluralize(filter.documents.total_count),
-    departments: filter_results_selections(filter.selected_organisations, 'departments'),
-    topics: filter_results_selections(filter.selected_topics, 'topics'),
-    taxons: filter_taxon_selections(
-      filter.selected_taxons,
-      filter.selected_subtaxons
-    ),
-    people: filter_results_selections(filter.selected_people_option, 'people'),
-    world_locations_any?: filter.selected_locations.any?,
-    world_locations: filter_results_selections(filter.selected_locations, 'world_locations'),
-    keywords: filter_results_keywords(filter.keywords),
-    date_from: filter.date_from,
-    date_to: filter.date_to,
-    include_world_location_news: filter.include_world_location_news,
-    filtering_command_and_act_papers?: filter.filtering_command_and_act_papers?,
-    filtering_command_papers_only?: filter.filtering_command_papers_only?,
-    filtering_act_papers_only?: filter.filtering_act_papers_only?
-  }
-%>
-<div class="filter-results-summary">
-  <%= render_mustache('documents/filter_selections', context) %>
-</div>
-
-<%# This will be hoisted into <head> by Slimmer %>
-<% tag :meta, name: 'robots', content: 'noindex' unless filter.documents.first_page? %>
