+++ app/views/publications/index.html.erb
@@ -19,23 +19,11 @@
   <%= render partial: "documents/filter_form",
              locals: {
                document_type: :publication,
-               filters: if Locale.current.english?
-                          [
-                            :keyword, :date, :publication_type, :locations,
-                            :department, :official_document_status, :taxon, :subtaxon
-                          ]
-                        else
-                          [ :locations ]
-                        end
+               filters: [ :locations ],
              } %>
 
   <div class="filter-results-block">
     <div class="inner-block">
-      <% if Locale.current.english? %>
-        <%= render partial: "documents/filter_results",
-                  locals: { filter: @filter,
-                            result_type: "publication" } %>
-      <% end %>
       <%= render 'shared/feeds', atom_url: filter_atom_feed_url %>
       <div class="filter-results js-filter-results" aria-live="polite">
         <%= render_mustache('documents/filter_table', @filter.as_hash) %>
