+++ app/views/statistics/index.html.erb
@@ -22,22 +22,11 @@
   <%= render partial: "documents/filter_form",
              locals: {
                document_type: :statistic,
-               filters: if Locale.current.english?
-                          [
-                            :keyword, :date, :department, :taxon
-                          ]
-                        else
-                          [ ]
-                        end
+               filters: [],
              } %>
 
   <div class="filter-results-block">
     <div class="inner-block">
-      <% if Locale.current.english? %>
-        <%= render partial: "documents/filter_results",
-                  locals: { filter: @filter,
-                            result_type: "statistics" } %>
-      <% end %>
       <%= render 'shared/feeds', atom_url: filter_atom_feed_url %>
       <div class="filter-results js-filter-results" aria-live="polite">
         <%= render_mustache('documents/filter_table', @filter.as_hash) %>
